dianhua/requirements.txt
aniso8601==1.2.0
appdirs==1.4.3
appnope==0.1.0
attrs==16.3.0
backports.shutil-get-terminal-size==1.0.0
#BeautifulSoup==3.2.1
cffi==1.9.1
click==6.6
clime==0.3.1
constantly==15.1.0
cryptography==1.7.1
cssselect==1.0.0
decorator==4.0.10
enum34==1.1.6
fake-useragent==0.1.3
Flask==0.10.1
Flask-RESTful==0.3.5
gevent==1.1.2
greenlet==0.4.10
gunicorn==19.6.0
hiredis==0.2.0
idna==2.1
incremental==16.10.1
ipaddress==1.0.17
#ipython==5.1.0
#ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.4
MarkupSafe==0.23
meld3==1.0.2
#mongolog==0.1.1
mrq==0.1.18
numpy==1.12.1
objgraph==1.8.1
packaging==16.8
pandas==0.19.2
parsel==1.1.0
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
pprintpp==0.2.3
prompt-toolkit==1.0.9
psutil==1.2.1
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
pycrypto==2.6.1
PyDispatcher==2.0.5
PyExecJS==1.4.0
Pygments==2.1.3
pymongo==3.3.0
pyOpenSSL==16.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.7
queuelib==1.4.2
redis==2.10.5
requests==2.12.1
rq==0.6.0
Scrapy==1.2.2
#selenium==3.0.1
service-identity==16.0.0
simplegeneric==0.8.1
six==1.10.0
subprocess32==3.2.5
supervisor==3.0
termcolor==1.1.0
traitlets==4.3.1
Twisted==16.6.0
ujson==1.33
uWSGI==2.0.14
w3lib==1.16.0
wcwidth==0.1.7
Werkzeug==0.11.11
xlrd==1.1.0
zope.interface==4.3.3
tornado==5.0.2
